implements method that removes tasks from db

+++ TaskerProject/Tasker/Actions/TaskAction.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Data.Entity;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -9,13 +10,13 @@ namespace Tasker
 	public class TaskAction
 	{
 		DataBaseContext db;
-		
+
 		public void Add(TaskClass currentTaskClass)
 		{
 			using (db = new DataBaseContext())
 			{
-					if (currentTaskClass == null)
-						throw new Exception("You are tring to add Task that is null");
+				if (currentTaskClass == null)
+					throw new Exception("You are tring to add Task that is null");
 				db.TaskSet.Add(currentTaskClass);
 				db.SaveChanges();
 			}
@@ -37,12 +38,15 @@ namespace Tasker
 		{
 			using (db = new DataBaseContext())
 			{
-				TaskClass currentTask = new TaskClass(task);
+				TaskClass currentTask;
+				var taskToDeleteCollection = db.TaskSet.Where(x => x.Title == task).ToList();
+				if (taskToDeleteCollection.Count > 1)
+					throw new Exception("There are two records with the same name");
+				currentTask = taskToDeleteCollection.FirstOrDefault() as TaskClass;
 				db.TaskSet.Remove(currentTask);
 				db.SaveChanges();
 			}
-
-
 		}
 	}
 }
+
